Move pytorch model tests up a directory, removing tank/pytorch/.
Fix formatting.

fixup formatting.
