+++ tank/tf/hf_masked_lm/funnel-transformer_test.py
@@ -92,19 +92,25 @@ class FunnelModuleTest(unittest.TestCase):
         self.module_tester.save_vmfb = pytestconfig.getoption("save_vmfb")
         self.module_tester.benchmark = pytestconfig.getoption("benchmark")
 
-    @pytest.mark.skip(reason="Funnel currently fails during the lowering passes")
+    @pytest.mark.skip(
+        reason="Funnel currently fails during the lowering passes"
+    )
     def test_module_static_cpu(self):
         dynamic = False
         device = "cpu"
         self.module_tester.create_and_check_module(dynamic, device)
 
-    @pytest.mark.skip(reason="Funnel currently fails during the lowering passes")
+    @pytest.mark.skip(
+        reason="Funnel currently fails during the lowering passes"
+    )
     def test_module_dynamic_cpu(self):
         dynamic = True
         device = "cpu"
         self.module_tester.create_and_check_module(dynamic, device)
 
-    @pytest.mark.skip(reason="Funnel currently fails during the lowering passes")
+    @pytest.mark.skip(
+        reason="Funnel currently fails during the lowering passes"
+    )
     @pytest.mark.skipif(
         check_device_drivers("gpu"), reason="nvidia-smi not found"
     )
@@ -113,7 +119,9 @@ class FunnelModuleTest(unittest.TestCase):
         device = "gpu"
         self.module_tester.create_and_check_module(dynamic, device)
 
-    @pytest.mark.skip(reason="Funnel currently fails during the lowering passes")
+    @pytest.mark.skip(
+        reason="Funnel currently fails during the lowering passes"
+    )
     @pytest.mark.skipif(
         check_device_drivers("gpu"), reason="nvidia-smi not found"
     )
@@ -122,7 +130,9 @@ class FunnelModuleTest(unittest.TestCase):
         device = "gpu"
         self.module_tester.create_and_check_module(dynamic, device)
 
-    @pytest.mark.skip(reason="Funnel currently fails during the lowering passes")
+    @pytest.mark.skip(
+        reason="Funnel currently fails during the lowering passes"
+    )
     @pytest.mark.skipif(
         check_device_drivers("vulkan"),
         reason="vulkaninfo not found, install from https://github.com/KhronosGroup/MoltenVK/releases",
@@ -132,7 +142,9 @@ class FunnelModuleTest(unittest.TestCase):
         device = "vulkan"
         self.module_tester.create_and_check_module(dynamic, device)
 
-    @pytest.mark.skip(reason="Funnel currently fails during the lowering passes.")
+    @pytest.mark.skip(
+        reason="Funnel currently fails during the lowering passes."
+    )
     @pytest.mark.skipif(
         check_device_drivers("vulkan"),
         reason="vulkaninfo not found, install from https://github.com/KhronosGroup/MoltenVK/releases",
